+++ CHANGELOG.md
@@ -7,10 +7,16 @@
 * All fields of `Composition` are now private (!273)
 * `Input.source` is private, `Input.get_source()` becomes property `Input.source`, `Input.set_source` becomes a getter (!273)
 * `Output.targets` is private, `Output.get_targets()` becomes property `Output.targets` (!273)
+* Composition metadata was restructured to hold components and adapters in separate sub-dictionaries (!274)
+
+### Features
+
+* Components and adapters automatically provide default metadata that can be extended by implementations (!274)
 
 ### Documentation
 
 * Minor fixes in documentation examples and links (!272)
+* Adds a book section on composition, component and adapter metadata (!274)
 
 ## [v0.5.1]